+++ src/main/java/com/codeborne/selenide/commands/SetValue.java
@@ -67,10 +67,10 @@ public class SetValue implements Command<WebElement> {
             "if (webelement.getAttribute('disabled') != undefined) return 'Cannot change value of disabled element';" +
             "webelement.focus();" +
             "var maxlength = webelement.getAttribute('maxlength') == null ? -1 : parseInt(webelement.getAttribute('maxlength'));" +
-            "webelement.value = " +
+            "webelement.setAttribute('value', " +
             "maxlength == -1 ? text " +
             ": text.length <= maxlength ? text " +
-            ": text.substring(0, maxlength);" +
+            ": text.substring(0, maxlength));" +
             "return null;" +
             "})(arguments[0], arguments[1]);",
         element, text);