+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.8.0 (released 28.02.2020)
 * #662 Added Condition with Predicate<WebElement> parameter as alternative to check element conditions  --  thanks to Dmytro Stekanov for #1059
 * #845 Added methods $.sibling() and $.preceding()  --  thanks to Dmytro Stekanov for #1064
+* #994 Added method to check content of ":before" and other pseudo elements  --  thanks to Denys Shynkarenko for PR #1045
 * #1071 Fixing issue when SoftAssertionsExtension for JUnit5 sometimes marked passed tests as failed 
 * $.click() now clicks the CENTER of element (also for Configuration.clickViaJS=true)